+++ tests/integration/defs/test_e2e.py
@@ -1314,6 +1314,13 @@ def test_openai_chat_example(llm_root, llm_venv):
     llm_venv.run_cmd(["-m", "pytest", str(test_root / "_test_openai_chat.py")])
 
 
+def test_openai_reasoning(llm_root, llm_venv):
+    test_root = unittest_path() / "llmapi" / "apps"
+    llm_venv.run_cmd(
+        ["-m", "pytest",
+         str(test_root / "_test_openai_reasoning.py")])
+
+
 def test_openai_chat_multimodal_example(llm_root, llm_venv):
     example_root = Path(os.path.join(llm_root, "examples", "apps"))
     test_root = unittest_path() / "llmapi" / "apps"
