+++ tests/integration/test_lists/test-db/l0_a10.yml
@@ -74,6 +74,7 @@ l0_a10:
   - test_e2e.py::test_openai_misc_example
   - test_e2e.py::test_openai_completions_example
   - test_e2e.py::test_openai_chat_example
+  - test_e2e.py::test_openai_reasoning
   - test_e2e.py::test_openai_chat_multimodal_example
   - test_e2e.py::test_trtllm_bench_sanity[non-streaming-FP16-meta-llama/Llama-3.1-8B-llama-3.1-model/Meta-Llama-3.1-8B]
   - test_e2e.py::test_trtllm_bench_latency_sanity[FP16-meta-llama/Llama-3.1-8B-llama-3.1-model/Meta-Llama-3.1-8B]
